requirements.txt
mysql-connector-python~=8.0.22
SQLAlchemy~=1.3.20